notification list

+++ jobs/company/views.py
@@ -1,8 +1,3 @@
-from functools import partial
-from inspect import stack
-from django.shortcuts import render
-from requests import request
-from rest_framework import generics
 from rest_framework.views import APIView
 from .serializers import JobSerializer, JobUpdateSerializer, JobCreateSerializer
 from jobs.models import Job
@@ -13,15 +8,8 @@ from django.http import Http404
 from rest_framework.decorators import api_view
 
 
-# class ListCompanyJobs(generics.ListAPIView):
-#     #self.request.user
-#     user=User.objects.get(id=5)
-#     queryset=Job.objects.filter(job_owner=user)
-#     serializer_class=JobSerializer
 class ListCompanyJobs(APIView):
-    # authentication_classes = [authentication.TokenAuthentication]
-    # permission_classes = [permissions.IsAdminUser]
-
+   
     def get(self, request, format=None):
         """
         Return a list of all jobs related to the requested user.
@@ -46,35 +34,52 @@ class ListCompanyJobs(APIView):
 
 
 class RetrieveUpdateDeleteCompanyJob(APIView):
+    """
+    get the job object 
+    """
     def get_object(self, pk):
         try:
             return Job.objects.get(pk=pk)
         except Job.DoesNotExist:
             raise Http404
-
+    """
+    retrieve one job
+    """
     def get(self, request, pk, format=None):
+        print(request.user)
         job = self.get_object(pk)
         serializer = JobSerializer(job)
-        print(serializer)
         return Response(serializer.data)
-
+    """
+    Update job only if it's open and has permission to update it as a job owner 
+    """
     def put(self, request, pk, format=None):
+        
         job = self.get_object(pk)
-        serializer = JobUpdateSerializer(job, data=request.data, partial=True)
-        if serializer.is_valid():
-            if job.status == "OPEN":
-                serializer.save()
-                return Response(serializer.data)
+        if request.user == job.job_owner:
+            serializer = JobUpdateSerializer(job, data=request.data, partial=True)
+            if serializer.is_valid():
+                if job.status == "OPEN":
+                    serializer.save()
+                    return Response(serializer.data)
+        else:
+            return Response({"details": "You don't have the persmission to edit or delete this job"}, status=status.HTTP_403_FORBIDDEN)
         return Response({"details": "your job cannot be edited "}, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, pk, format=None):
         job = self.get_object(pk)
         st = status.HTTP_400_BAD_REQUEST
-        if job.status == 'OPEN':
-            print('got delete')
-            st = status.HTTP_204_NO_CONTENT
-            job.delete()
-            return Response({"details": "your job is deleted"}, status=st)
+        if request.user == job.job_owner:
+            if job.status == 'OPEN':
+                print('got delete')
+                st = status.HTTP_204_NO_CONTENT
+                job.delete()
+                return Response({"details": "your job is deleted"}, status=st)
+        else:
+            return Response({"details": "You don't have the persmission to edit or delete this job"}, status=status.HTTP_403_FORBIDDEN)
+        return Response({"details": "job is not deleted"}, status=status.HTTP_400_BAD_REQUEST)
+
+
 
 
 @api_view(['POST'])
@@ -83,8 +88,9 @@ def AcceptDeveloper(request, pk):
     job = Job.objects.get(pk=pk)
     id = request.data['id']
     user = User.objects.get(id=id)
-    if job.status == 'OPEN' and job.developer is None:
+    if job.status == 'OPEN' and job.developer is None and request.user==job.job_owner:
         job.developer=user
         job.save()
         return Response({"details": "Developer has been accepted"}, status=status.HTTP_201_CREATED)
+    
     return Response({"details": "Developer can't be accepted"}, status=status.HTTP_201_CREATED)
